Move getters/setters to utility functions
