+++ packages/marko-web-omeda-identity-x/utils/get-answered-question-map.js
@@ -1,0 +1,15 @@
+/**
+ * Builds a map of user answers to custom select/boolean questions.
+ */
+module.exports = (user) => {
+  const answeredQuestionMap = new Map();
+  user.customSelectFieldAnswers.forEach((select) => {
+    if (!select.hasAnswered) return;
+    answeredQuestionMap.set(select.field.id, true);
+  });
+  user.customBooleanFieldAnswers.forEach((boolean) => {
+    if (!boolean.hasAnswered) return;
+    answeredQuestionMap.set(boolean.field.id, true);
+  });
+  return answeredQuestionMap;
+};
